pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
    "setuptools-scm[toml]>=6.2",
    "setuptools>=45",
]

[project]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python",
    "Topic :: Software Development",
    "Typing :: Typed",
]
dependencies = [
    "PyYAML",
    "jax",
    "jaxlib",
    "matplotlib",
    "py-cpuinfo",
    "tqdm",
]
description = "Set of commands for benchmarking JAX on your machine"
dynamic = ["version"]
keywords = [
    "HEP",
    "PWA",
    "amplitude analysis",
    "partial wave analysis",
    "particle physics",
    "physics",
    "tensorflow",
]
license = { file = "LICENSE" }
maintainers = [
    { name = "Common Partial Wave Analysis", email = "compwa-admin@ep1.rub.de" },
]
name = "jax-mini-benchmark"
readme = "README.md"
requires-python = ">=3.7"

[project.optional-dependencies]
dev = [
    "jax-mini-benchmark[sty]",
]
format = [
    "black",
    "isort",
]
sty = [
    "jax-mini-benchmark[format]",
    "pre-commit",
]

[project.scripts]
benchmark-jax = "jax_benchmark:main"
benchmark-jax-dot-product = "jax_benchmark.dot_product:main"

[project.urls]
Documentation = "https://github.com/ComPWA/jax-mini-benchmark"
Source = "https://github.com/ComPWA/jax-mini-benchmark"
Tracker = "https://github.com/ComPWA/jax-mini-benchmark/issues"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
jax_benchmark = ["py.typed"]


[tool.setuptools_scm]
write_to = "src/jax_benchmark/_version.py"
